LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/origin-keyed-agent-clusters/getter-special-cases/cross-origin-isolated.sub.https-expected.txt

Blocked access to external URL https://www1.localhost:9443/html/browsers/origin/origin-keyed-agent-clusters/resources/coep-frame.html


Harness Error (TIMEOUT), message = null

FAIL self: originAgentCluster must equal true assert_equals: expected (boolean) true but got (undefined) undefined
TIMEOUT child: originAgentCluster must equal true Test timed out
